Start of new tests
I have no clue how I am going to do this. I do not want to write a
test for this test. I have to write it in such a way that the
validity of the code is not in question. I need to be clever with
the properties. I have no clue how I am going to do this.