Setup of variables n loop
I do not like that the number of properties has to be hardcoded
like that. Because if I add a property I have to remember to
increment that variable, which is tedious. But it is easy to
understand, so I guess it is ok.

Do slices of type bool default to false? I hope so.