It doesn't make sense
I just realized that on line 34 I declared a slice, not an array.
That means that I shouldn't be able to use an index ... and yet
the compiler is letting me. What is going on in here? Does this
means that the index gets populated automagically (like in
JavaScript)? I hope not.

I was thinking that I could need to use append, which would be a
nigthmare because the index would not be meaningful. No, I need to
initialize it.

Can I give the slice a size? The numOfProperties is a constant. So
I guess I can use it right? I recall not being able to use a const
before. I don't remember. I will give it a try.