Test prop 2
I am using an online editor to test the regular expresion. There
is an issue though: I need to make sure that the entore string
matches this pattern parfectly. That is, it must match it exacly
once.

A big issue that I have so far with my experience with Go is that
testing things takes a lot of time. Repls spoiled me too much. I
was reading the docs and I think that to impl the functionality
that I need I have to use FindAllString and check that the value
is 1. But first I need to make sure I understand what the function
is doing.