This is not going to work
As I was working on the reg exps I started to think about the
types of strings that I want to test that my program can handle.
In particular, I realized that the term "([^ ]+ )" in the regex
has only one whitespace. The regex I indicated accepts at least one
white space in the front, zero or more symbols followed by a single
whitespace, and at least one symbol at the end. That is strings
like this:

\w -> whitespace(for clarity when two or more are used)
\0 -> end of string

ACCEPT:
\w\w\w\w\wThis string is accepted.\0
\w\wThis is also accepted\0

REJECT:
\w\wBut\w\wthis would be\w\w\wrejected.\0

And I realized that this rejected case is something that I want to
take into account too. You might be asking, why the obscetion with
whitespaces? Well, whitespaces are crucial in spliting strings into
"words". It is really easy to make mistakes around the usage of \w
to split strings into "words". I am betting that there will be bugs
around this char. Thus, I want to create as many edge cases as
possible.

This means that I have to come up with more cases. As I was looking
at the complexity of these patterns, I realized that the approach
that I am taking in flawed. The chances of it giving a false
positive are rather high. For example I was thinking of checking
for both cases: " +([^ ]+ )*[^ ]+" and " +([^ ]+ +)*[^ ]+". The
first is the original regex, the second is the one that accepts
the rejected example above. I think [^ ]+ + <==> .+ but I prefer
the former since it is more explitic. I think it is easier to
understand. I was planning on testing for both. I was thinking
about the current impl and how this feature would need to be impl
by generating a random number to decide how many whitepsaces to
have between words. What are the odds that the generator would
produce single whitespaces for the entire string? Low. Even if I
am running the program 10000 times, the more cases I add, the more
likely it is for any one of them to fail. This strategy does not
provide confidence. More than anything, what it provides are hard,
logic/probabilistic bugs and a lot of work. This is definetely not
worth it.

It is possible to describe a stochastic process as nondeterministic
merge of deterministic operations. That is what I originally
thought and planned with a stochastic builder pattern. I think that
is the way to go. i need to review that.

But definetely not this.